Implement a RC schema for the package
So RC version of Xen can be build later, there should be no change for
the package in this patch.

In CentOS (or any RPM based package), to build a package for a release
candidate, the Version field in the package is the version that is going
to be released, and the Release field of the package contains the RC
number and start by "0.".

e.g., to build a package for for Xen 4.10 RC8, the package version shown
in yum would be "4.10.0-0.1.rc8.el7.centos", for the next RC, that could
be "4.10.0-0.2.rc9.el7.centos".