lib: find-duplicate-patch-ids: Remove duplicates in upstream_pids
It may happen that a patch is found twice in git. (e.g. between
RELEASE-4.10.0 and RELEASE-4.10.1), so remove the duplicates patch-ids
before finding duplicates between upstream_pids and series_pids.

This change the first value in the lists of pids to be the same for
every pids that comes from git. Those values are just commit hash we
don't need. The choosen value "0000_upstream" is so that the `sort -k2
-r` later will list that commit after the equivalent patch (same
patch-id) from stg.

Signed-off-by: Anthony PERARD <anthony.perard@citrix.com>