4.8.4.42.gd95b5bb31e-1: Build from staging tree to include XSA 268-273
XSA 273 "patches" are actually a commit id from the xen git tree. Use it
to build the package.

The package version and tarball name generated by xen.git are different,
we cann't have '-' in an rpm version, so change the SPEC to accomodate
the discrepancy.

Another issue, the binary of xen in /boot doesn't match the version
either, so rename it.