lib: Generate a tarball Xen from a git tree
Handle the creation of a tarball from a Xen commit id. They are
verification in place to check that the commit contain the tag of the
release we want to build.

We make use of `make src-tarball` without running ./configure first so
its run is faked by `touch config/Tools.mk`. Still, that not enough, the
system that will run make src-tarball will needs at least gcc.