4.10.1.106.g13e85a6dbc-2: Fix the Xen version string
The version, as known by Xen was 4.10.2-pre-1.el7 after recent changes. Define
XEN_EXTRAVERSION so that it will be the same as the package version.

Also remove the temporary git repo so Xen build system will no try to extract
information from it.

Also, fix the %prep phase, as make -C xen olddefconfig is called there, it
would generate a /boot/xen-*.config with a bogus xen version in it. Set
XEN_EXTRAVERSION earlier.