lib, get-xen-stable: Improve handling of git snapshot and version
With this patch, it is possible to only update XEN_CSET in sources.cfg,
then the function get-xen-stable uses XEN_CSET and information from
xen.spec (nb_commit) to generate the name for the snapshot tarball. If
that tarball doesn't exist it simply generate a new one and correct the
information in the SPEC (nb_commit and abbrev_cset).

In the SPEC, nb_commit and abbrev_cset are information about a specific
commit id that are necessary to use a snapshot (git describe provide
those information in a string). The SPEC will use those two values and
the xen_version to generate the Version of the package as well as the
tarball name of the snapshot in order to be able to extract it.

When a commit id correspond to an release tag, nb_commit and abbrev_cset
aren't actually used and will not be included in the package version.

Signed-off-by: Anthony PERARD <anthony.perard@citrix.com>